Move canvas tab hit tester to own file
