+++ Packages/macOS/CmuxCanvasUI/Sources/CmuxCanvasUI/CanvasPaneTitleBarView.swift
@@ -10,25 +10,6 @@ struct CanvasTabHitRegions: Equatable {
     var closeFrames: [UUID: CGRect] = [:]
 }
 
-struct CanvasTabHitTester {
-    var tabOrder: [UUID]
-    var hitRegions: CanvasTabHitRegions
-
-    func tab(at point: CGPoint) -> UUID? {
-        tabOrder.first { tabId in
-            hitRegions.tabFrames[tabId]?.contains(point) == true
-        }
-    }
-
-    func closeTab(at point: CGPoint, hoveredTabId: UUID?) -> UUID? {
-        guard let hoveredTabId,
-              hitRegions.closeFrames[hoveredTabId]?.contains(point) == true else {
-            return nil
-        }
-        return hoveredTabId
-    }
-}
-
 private struct CanvasTabFramesKey: PreferenceKey {
     static let defaultValue = CanvasTabHitRegions()
     static func reduce(value: inout CanvasTabHitRegions, nextValue: () -> CanvasTabHitRegions) {
